+++ app/src/main/AndroidManifest.xml
@@ -10,14 +10,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".controller.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <receiver android:name="C19WidgetProvider">
+        <receiver android:name=".view.C19WidgetProvider">
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
             </intent-filter>
@@ -25,7 +25,7 @@
                 android:name="android.appwidget.provider"
                 android:resource="@xml/info_appwidget" />
         </receiver>
-        <service android:name=".C19WidgetProvider$UpdateService" />
+        <service android:name=".view.C19WidgetProvider$UpdateService" />
     </application>
 
 </manifest>
