+++ app/src/main/java/com/android/c19widget/controller/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.android.c19widget
+package com.android.c19widget.controller
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.android.c19widget.R
 
 class MainActivity : AppCompatActivity() {
