+++ app/src/main/java/com/android/c19widget/model/Covid9jaApi.kt
@@ -1,4 +1,4 @@
-package com.android.c19widget.data
+package com.android.c19widget.model
 
 import retrofit2.Call
 import retrofit2.http.GET
