+++ app/src/main/java/com/android/c19widget/model/Covid9jaApiResponse.kt
@@ -1,3 +1,3 @@
-package com.android.c19widget.data
+package com.android.c19widget.model
 
 data class Covid9jaApiResponse(val categories: String, val values: String)
