+++ app/src/main/java/com/android/c19widget/model/Repository.kt
@@ -1,4 +1,4 @@
-package com.android.c19widget.data
+package com.android.c19widget.model
 
 interface Repository {
